refactored email marketing agent
